Server empfängt die Länge von Tcp-Packeten jetzt Richtig

+++ src/de/_13ducks/spacebatz/client/network/ClientNetwork.java
@@ -86,7 +86,7 @@ public class ClientNetwork {
 
 
             // Packetlänge senden
-            sendStream.writeLong(message.length);
+            sendStream.writeShort(message.length);
 
             // alle Hunderterblöcke senden:
             byte msg[] = new byte[100];
